Remove conda from CI and setup scripts (#774)
This PR removes conda environments from the Dockerfiles and setup
scripts.

For the CI, we use a vanilla virtualenv to ensure that the garage
package installs in the lowest common-denonenator environment. We also
add a job to TravisCI to verify that conda can successfully build.

For the setup scripts, we remove the environment creation altogether and
assume that the user will install garage in her own environment of
choice. The scripts serve to only setup pre-requisities, not the Python
environment itself

Summary of changes:
* Remove conda environment from Dockerfiles and setup scripts
* Add virtualenv to Dockerfiles
* Add TravisCI job to build verify conda and pipenv can install garage
* Add pytest-xdist for parallel tests
* Update some tests and code to be Python 3.5 compliant
* Update installation documentation